requirements.txt
Flask==3.0.3
pillow==10.3.0
python-tgpt==0.5.3
requests==2.28.2